# openapi/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Spring Application exposes Api to manage configuration for EC/PSP on the Nodo dei Pagamenti",
    "termsOfService": "https://www.pagopa.gov.it/",
    "title": "PagoPA API configuration",
    "version": "0.2.28"
  },
  "host": "127.0.0.1:8080",
  "basePath": "/apiconfig/api/v1",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/counterparttables": {
      "get": {
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "default": 50,
            "description": "Number of elements on one page. Default = 50",
            "format": "int32",
            "in": "query",
            "name": "limit",
            "required": false,
            "type": "integer"
          },
          {
            "description": "Page number. Page value starts from 0",
            "format": "int32",
            "in": "query",
            "name": "page",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/CounterpartTables"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemJson"
            }
          },
          "500": {
            "description": "Service unavailable",
            "schema": {
              "$ref": "#/definitions/ProblemJson"
            }
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "Authorization": []
          }
        ],
        "tags": [
          "Creditor Institutions"
        ],
        "operationId": "getCounterpartTables",
        "summary": "Get the counterparties table"
      }
    },
    "/counterparttables/{idcounterparttable}": {
      "get": {
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "description": "Id counterpart table",
            "in": "path",
            "name": "idcounterparttable",
            "required": true,
            "type": "string"
          },
          {
            "description": "Creditor institution code",
            "in": "query",
            "name": "creditorinstitutioncode",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "format": "binary",
              "type": "string"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemJson"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemJson"
            }
          },
          "500": {
            "description": "Service unavailable",
            "schema": {
              "$ref": "#/definitions/ProblemJson"
            }
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "Authorization": []
          }
        ],
        "tags": [
          "Creditor Institutions"
        ],
        "operationId": "getCounterpartTable",
        "summary": "Download a XML file containing the details of a counterpart table"
      }
    },
...
  },
  "definitions": {
...
    "CounterpartTable": {
      "properties": {
        "business_name": {
          "example": "Comune di Lorem Ipsum",
          "type": "string"
        },
        "creditor_institution_code": {
          "example": "1234567890100",
          "type": "string"
        },
        "id_counterpart_table": {
          "example": "123456789",
          "type": "string"
        },
        "publication_date": {
          "format": "date-time",
          "type": "string"
        },
        "validity_date": {
          "format": "date-time",
          "type": "string"
        }
      },
      "required": [
        "business_name",
        "creditor_institution_code",
        "id_counterpart_table",
        "publication_date",
        "validity_date"
      ],
      "type": "object"
    },
    "CounterpartTables": {
      "properties": {
        "counterpart_tables": {
          "items": {
            "$ref": "#/definitions/CounterpartTable"
          },
          "type": "array"
        },
        "page_info": {
          "$ref": "#/definitions/PageInfo"
        }
      },
      "required": [
        "counterpart_tables",
        "page_info"
      ],
      "type": "object"
    },
...
    "PageInfo": {
      "properties": {
        "items_found": {
          "description": "Number of items found. (The last page may have fewer elements than required)",
          "format": "int32",
          "type": "integer"
        },
        "limit": {
          "description": "Required number of items per page",
          "format": "int32",
          "type": "integer"
        },
        "page": {
          "description": "Page number",
          "format": "int32",
          "type": "integer"
        },
        "total_pages": {
          "description": "Total number of pages",
          "format": "int32",
          "type": "integer"
        }
      },
      "required": [
        "items_found",
        "limit",
        "page",
        "total_pages"
      ],
      "type": "object"
    },
...
    "ProblemJson": {
      "properties": {
        "detail": {
          "description": "A human readable explanation specific to this occurrence of the problem.",
          "example": "There was an error processing the request",
          "type": "string"
        },
        "status": {
          "description": "The HTTP status code generated by the origin server for this occurrence of the problem.",
          "example": 200,
          "format": "int32",
          "maximum": 600,
          "minimum": 100,
          "type": "integer"
        },
        "title": {
          "description": "A short, summary of the problem type. Written in english and readable for engineers (usually not suited for non technical stakeholders and not localized); example: Service Unavailable",
          "type": "string"
        }
      },
      "type": "object"
    },
...
  },
  "securityDefinitions": {
    "ApiKey": {
      "description": "The API key to access this function app.",
      "in": "header",
      "name": "Ocp-Apim-Subscription-Key",
      "type": "apiKey"
    },
    "Authorization": {
      "bearerFormat": "JWT",
      "description": "JWT token get after Azure Login",
      "scheme": "Bearer",
      "type": "apiKey",
      "in": "header",
      "name": "Authorization"
    }
  },
...
}